add experimental use of sqlite for data caching

+++ src/main/java/net/querz/mcaselector/io/mca/ChunkData.java
@@ -64,7 +64,7 @@ public class ChunkData {
 		}
 	}
 
-	public long parseData(OverlayDataParser parser) {
+	public int parseData(OverlayDataParser parser) {
 		return parser.parseValue(this);
 	}
 }
